+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.6.0
+- IsSecretConfigurationKey added to StartupConfiguration
+- ILogger and ILoggerFactory added to IBuildContext
+- Evaluators can be set to ConfigurationBuilder.Properties for usin in include blocks
+
+
 # 1.5.0
 - Limited support of placeholders for ${include}
 
@@ -12,7 +18,7 @@
 - Property evaluation fixed
 - IPropertyEvaluator interface changed
 
-# 1.1.0 - 15.05.2017
+# 1.1.0 - 15.05.2020
 - All merged to one package MicroElements
 - Microsoft.Extensions updated to 3.1.0